+++ public/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Pre-release v4.1.1-beta.5 - Unreleased
 * **Compatibility with Autosubmit API version >= [v4.1.2b7](https://pypi.org/project/autosubmit-api/v4.1.2b7/)** (unreleased)
+* Now the Quick View supports pagination and lazy loading.
 * Visual improvements (invert arrow direction in drop-down sections)
 
 ### Pre-release v4.1.1-beta.4 - Release date: 2026-08-11